+++ src/resources/collections/get.ts
@@ -76,9 +76,17 @@ export default class Get extends AbstractGet {
    */
   public async recursive(resource, opts, args) {
     for(let child of this.children) {
-      args.collection = resource.name;
+      let requested = child.names.filter(value => -1 !== opts.whitelist.indexOf(value));
+      let newArgs =  Object.assign({}, args);
+
+      newArgs.collection = resource.name;
       var instance = new child.resource(this.api);
-      instance.execute(opts, args, {});
+
+      if(requested.length === 0 && instance.getChildren().length === 0) {
+        continue;
+      }
+
+      instance.execute(opts, newArgs, {});
     }
   }
 }